feat: added delete image button

+++ menu-api/blueprints/menus/controllers/menus_resources.py
@@ -59,7 +59,7 @@ class AllMenuResource(MenusBaseResource):
         page = args["page"]
         menus = [menu.slug for menu in Menu.objects()]
 
-        return {"menus": menus[(page - 1) * limit : page * limit]}
+        return {"menus": menus[(page - 1) * limit: page * limit]}
 
 
 @doc(description="""Upload menu to server""")
@@ -146,7 +146,7 @@ class ImportMenuResource(MenusBaseResource):
         self.all_sections = {}
 
 
-@doc(description="""Menu element related operations""",)
+@doc(description="""Menu element related operations""", )
 class MenuResource(MenusBaseResource):
     @marshal_with(GetMenuSchema)
     def get(self, slug):
@@ -282,10 +282,13 @@ class ImageMenuResource(MenusBaseResource):
 
         item = menu.get_item(item_id)
         if item:
-            delete_file(item.image)
-            item.image = None
-            menu.save()
-            return item
+            if item.image:
+                delete_file(item.image)
+                item.image = None
+                menu.save()
+                return item
+            else:
+                return {'description': 'image not found'}
 
         return {"description": "Item not found"}
 
@@ -401,7 +404,7 @@ class ItemMenuResource(MenusBaseResource):
         if section_id not in section_ids:
             return {"description": "Invalid section"}, 404
 
-        item = Item(_id=str(uuid.uuid4()), sections=[section_id],)
+        item = Item(_id=str(uuid.uuid4()), sections=[section_id], )
 
         menu.menu_items.append(item)
         menu.save()
